requirements.txt
datasets
docker
fastapi
pandas
pyvespa
openai
python-dotenv
tqdm
